POC: evals metadata in builtin JUnit reporter
These changes show how we can use test.annotations to inject the evals data
in the built-in JUnit reporter (no changes needed)

The annoying bit is that we can only use the `<property>` tags and have a message
in them. So we relied on cheap naming tricks to make sense of the messages.

That's not amazing, and would be not ideal to stuff the rationale, for example.
But it's doable none the less.

To really use this we can have the TestAnalytics JUnit parser look for these properties
that start with `evals.` and parse the name into something meaningful.